+++ db/upgrade.php
@@ -151,6 +151,12 @@ function xmldb_format_grid_upgrade($oldversion = 0) {
         upgrade_plugin_savepoint(true, 2020111402, 'format', 'grid');
     }
 
+    if ($oldversion < 2020111403) {
+        // Any transparent PNG's need regenerating.
+        format_grid::update_displayed_images_callback();
+        upgrade_plugin_savepoint(true, 2020111403, 'format', 'grid');
+    }
+
     // Automatic 'Purge all caches'....
     purge_all_caches();